Add filter for the og:title
By default the plugin WordPress SEO by Yoast uses the title as created to be shown in the browser which includes the site name if the title is not given in the post under Social.
Facebook recommends to exclude branding.
Also create a more elaborate title for the dynamic pages.