Change the Sitemap data for the front page.
As WordPress SEO by Yoast does not change the data for the front page, we do this ourselves.
We set the priority to 100%, the Change Frequency to Daily, and set the last modified date by getting the articles on the front page and using the most recent modified date of that post, If we don't do this, the entry will take the modified date from the actual page, which never changes.